pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tasks-tui-app"
version = "0.1.7"
description = "A command-line TUI for managing Google Tasks."
readme = "README.md"
authors = [
    {name = "Huy Pham", email = "huyquangpham2005@gmai.com"},
]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Environment :: Console :: Curses",
]
dependencies = [
    "google-api-python-client",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "uni-curses",
    "python-dateutil",
    "textual",
]

[project.scripts]
# This line (tasks-tui = ...) is likely what was failing to parse.
tasks-tui = "tasks_tui.main:cli"
tasks-tui-textual = "tasks_tui.textual_ui.app:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["tasks_tui"]

